requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
pydantic
passlib
python-jose